requirements.txt
django
djangorestframework 